pcb_inspection/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
# anomalib>=1.0.0  # Removed - not compatible with Python 3.14
ultralytics>=8.1.0
onnxruntime>=1.16.0
opencv-python>=4.8.0
albumentations>=1.3.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.20
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
Pillow>=10.0.0
requests>=2.31.0
matplotlib>=3.7.0
streamlit>=1.28.0
lightning>=2.0.0
